ai_info_assistant/market_intel_ai/pyproject.toml
[project]
name = "market-intel-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
